requirements.txt
openai==0.28.0
fastapi
pymongo
python-dotenv
uvicorn
sentence_transformers
numpy